Prevent integer overflow - Bug #19895
git-svn-id: file:///svn/phpbb/trunk@8877 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/acp/acp_forums.php
@@ -1817,7 +1817,7 @@ class acp_forums
 		$row = $db->sql_fetchrow($result);
 		$db->sql_freeresult($result);
 
-		set_config('upload_dir_size', (int) $row['stat'], true);
+		set_config('upload_dir_size', (float) $row['stat'], true);
 
 		return array();
 	}
